Domain Reliability: More security review.
Change some trivial stuff so sleevi can do more security review.
Also, add rdsmith as an OWNER.

BUG=356791

Review URL: https://codereview.chromium.org/252613002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@269434 0039d316-1c4b-4281-b951-d872f2087c98

+++ components/domain_reliability/uploader_unittest.cc
@@ -30,13 +30,13 @@ class DomainReliabilityUploaderTest : public testing::Test {
         uploader_(DomainReliabilityUploader::Create(
             url_request_context_getter_)) {}
 
-  DomainReliabilityUploader::UploadCallback MakeUploadCallback(int index) {
+  DomainReliabilityUploader::UploadCallback MakeUploadCallback(size_t index) {
     return base::Bind(&DomainReliabilityUploaderTest::OnUploadComplete,
                       base::Unretained(this),
                       index);
   }
 
-  void OnUploadComplete(int index, bool success) {
+  void OnUploadComplete(size_t index, bool success) {
     EXPECT_FALSE(upload_complete_[index]);
     upload_complete_[index] = true;
     upload_successful_[index] = success;
@@ -47,8 +47,10 @@ class DomainReliabilityUploaderTest : public testing::Test {
   scoped_refptr<net::TestURLRequestContextGetter> url_request_context_getter_;
   scoped_ptr<DomainReliabilityUploader> uploader_;
 
-  std::map<int, bool> upload_complete_;
-  std::map<int, bool> upload_successful_;
+  // Whether the upload callback was called for a particular collector index.
+  std::map<size_t, bool> upload_complete_;
+  // Whether the upload to a particular collector was successful.
+  std::map<size_t, bool> upload_successful_;
 };
 
 TEST_F(DomainReliabilityUploaderTest, Create) {
